remove disk, trainer modules, move training loop to main
